requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-marshmallow==1.1.0
marshmallow-sqlalchemy==1.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
marshmallow==3.20.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9